release 0.17.0

+++ app/controllers/api/v1/studies_controller.rb
@@ -432,7 +432,7 @@ module Api
         begin
           # create a map of file extension to use for creating directory_listings of groups of 10+ files of the same type
           @file_extension_map = {}
-          workspace_files = Study.firecloud_client.execute_gcloud_method(:get_workspace_files, @study.firecloud_project, @study.firecloud_workspace)
+          workspace_files = Study.firecloud_client.execute_gcloud_method(:get_workspace_files, 0, @study.firecloud_project, @study.firecloud_workspace)
           # see process_workspace_bucket_files in private methods for more details on syncing
           process_workspace_bucket_files(workspace_files)
           while workspace_files.next?
@@ -466,24 +466,37 @@ module Api
         # now remove files that we found that were supposed to be in directory_listings
         @unsynced_files.delete_if {|file| files_to_remove.include?(file.generation)}
 
-        # now check against latest list of files by directory vs. what was just found to see if we are missing anything and add directory to unsynced list
+        # now check against latest list of files by directory vs. what was just found to see if we are missing anything and
+        # add directory to unsynced list. also check if an existing directory is now 'orphaned' because it was moved and
+        # store that reference for removal after the check is complete
+        orphaned_directories = []
         @directories.each do |directory|
           synced = true
-          directory.files.each do |file|
-            if @files_by_dir[directory.name].detect {|f| f['generation'].to_s == file['generation'].to_s}.nil?
-              synced = false
-              directory.files.delete(file)
-            else
-              next
+          if @files_by_dir[directory.name].present?
+            directory.files.each do |file|
+              if @files_by_dir[directory.name].detect {|f| f['generation'].to_s == file['generation'].to_s}.nil?
+                synced = false
+                directory.files.delete(file)
+              else
+                next
+              end
             end
+            # if no longer synced, check if already in the list and remove as files list has changed
+            if !synced
+              @unsynced_directories.delete_if {|dir| dir.name == directory.name}
+              @unsynced_directories << directory
+            elsif directory.sync_status
+              @synced_directories << directory
+            end
+          else
+            # directory did not exist in @files_by_dir, so this directory_listing is orphaned and should be removed
+            orphaned_directories << directory
           end
-          # if no longer synced, check if already in the list and remove as files list has changed
-          if !synced
-            @unsynced_directories.delete_if {|dir| dir.name == directory.name}
-            @unsynced_directories << directory
-          elsif directory.sync_status
-            @synced_directories << directory
-          end
+        end
+
+        # remove orphaned directories
+        orphaned_directories.each do |orphan|
+          orphan.destroy
         end
 
         # provisionally save unsynced directories so we don't have to pass huge arrays of filenames/sizes in the form
@@ -492,6 +505,9 @@ module Api
           directory.save
         end
 
+        # reload unsynced directories to remove any that were orphaned
+        @unsynced_directories = @study.directory_listings.unsynced
+
         # split directories into primary data types and 'others'
         @unsynced_primary_data_dirs = @unsynced_directories.select {|dir| dir.file_type == 'fastq'}
         @unsynced_other_dirs = @unsynced_directories.select {|dir| dir.file_type != 'fastq'}
@@ -499,6 +515,10 @@ module Api
         # now determine if we have study_files that have been 'orphaned' (cannot find a corresponding bucket file)
         @orphaned_study_files = @study_files - @synced_study_files
         @available_files = @unsynced_files.map {|f| {name: f.name, generation: f.generation, size: f.upload_file_size}}
+
+        # now remove any 'bundled' files from @synced_study_files so we can render them inside their parent file's form
+        bundled_file_ids = @study.study_file_bundles.map {|bundle| bundle.bundled_files.to_a.map(&:id)}.flatten
+        @synced_study_files.delete_if {|file| bundled_file_ids.include?(file.id)}
       end
 
       private
